+++ src/applications/financial-status-report/pages/index.js
@@ -38,6 +38,7 @@ import * as utilityRecords from './expenses/utilities/records';
 import * as repayments from './expenses/repayments';
 import * as repaymentRecords from './expenses/repayments/records';
 import * as otherExpenses from './expenses/other';
+import * as otherExpenseRecords from './expenses/other/records';
 import * as resolutionOptions from './resolution/options';
 import * as resolutionComments from './resolution/comments';
 import * as bankruptcyHistory from './bankruptcy/history';
@@ -82,6 +83,7 @@ export {
   repayments,
   repaymentRecords,
   otherExpenses,
+  otherExpenseRecords,
   resolutionOptions,
   resolutionComments,
   bankruptcyHistory,
